+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+- Add [`nest`](https://github.com/Gabriella439/foldl/pull/215) for `Fold1`
+- Add [`Choice`, `Closed`, `Cosieve`, `Extend`, `Semigroupoid`, `Category`, `Strong`, `Arrow` and `ArrowChoice`](https://github.com/Gabriella439/foldl/pull/215) instances for `Fold1`
 - Add [`Closed`](https://github.com/Gabriella439/foldl/pull/215) instance for `Fold`
 
 1.4.17